backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
gunicorn
python-dotenv
ollama
requests
json
matplotlib
fastapi
openmeteo_requests
pytz
requests_cache
retry_requests